+++ R/tar_load.R
@@ -3,7 +3,7 @@
 #' @family storage
 #' @description Load the return values of targets into the current environment
 #'   (or the environment of your choosing). For a typical target, the return
-#'   value lives in a file in `_targets/objects/`. For dynamic files (i.e.
+#'   value lives in a file in `_targets/objects/`. For file targets (i.e.
 #'   `format = "file"`) the paths loaded in place of the values.
 #'   [tar_load_everything()] is shorthand for `tar_load(everything())`
 #'   to load all targets.