+++ R/tar_meta.R
@@ -82,7 +82,7 @@
 #'     on the result to locally recreate the target's initial RNG state.
 #'   * `path`: A list column of paths to target data. Usually, each element
 #'     is a single path, but there could be multiple paths per target
-#'     for dynamic files (i.e. `tar_target(format = "file")`).
+#'     for file targets (i.e. `tar_target(format = "file")`).
 #'   * `time`: `POSIXct` object with the time the target's data in storage
 #'     was last modified. If the target stores no local file,
 #'     then the time stamp corresponds to the time the target last